+++ vite.config.ts
@@ -21,5 +21,8 @@ export default defineConfig({
     port: 5173,
     strictPort: true,
     allowedHosts: true,
-  }
+  },
+  build: {
+    sourcemap: false,
+  },
 })